requirements.txt
flask==3.0.0
anthropic==0.39.0
python-dateutil==2.8.2
werkzeug==3.0.1
python-dotenv==1.0.0
